Change clickable components' cursor to pointer, add loading bar and make input textarea controlled
The button and clickable icon components will now have a pointer cursor. The button component now has an onClick property to which a function can be passed.
An animated loading bar has been added that will become visible when the submit button has been clicked.
The input textarea is now controlled, so that the contents can be used and, potentially, filtered.