Improve graph support and keys
The graph nows displays the data sent in by the backend, if no follow up is needed.
It also correctly parses the data.
The keys of dynamically generated elements are now based on their position in the respective array.
This is usually a bad practice, but since the positions are unique, minimal and identifiying and
because there are no other truely unique keys, it's okay.