Fix non-unique ids of graphs and bug related to redrawing them
The ids were non-unique because they all updated at the same time and thus their ids were all the same.
Furthermore, there was a bug related to redrawing them that caused the first graph to be emptied.
This was due to the querySelector just taking the first available svg, instead of checking for an id.
This has now been fixed.