Integrate priorConversation feature and saving settings
The interface will now remember the last graph uri from the X-qanary-graph
header and use it as a query parameter. If no graph uri is supplied,
no query parameter will be used in the next request.

Settings will now also be automatically saved in localStorage
when confirmed and restored at the start of each session.