Implement env var support
Environment variables are now used for configuration. The environment
variables are specified in the top-level directory's `.env` file.
They can be passed to a container on start-up to overwrite configuration
for images that were already built.

The documentation has been changed accordingly. Also, the "Before"
and "After" sections have been changed from ordered to unordered.